LibGfx/JPEGXL: Move the Channel utility class to its own file
